requirements.txt
streamlit
pandas
transformers
huggingface_hub
python-dotenv
torch